+++ src/sentry/workflow_engine/caches/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["CacheMapping"]
 
-from .cache_access import CacheMapping
+from .mapping import CacheMapping
